face_detection_project/requirements.txt
Flask==2.1.4
opencv-python==4.5.3.56
Pillow==8.3.2